backend/requirements.txt
click>=8.2
colorama>=0.4.6
Flask>=2.3
Flask-SQLAlchemy>=3.0.3
greenlet>=2.0.2
itsdangerous>=2.1.2
Jinja2>=3.1.2
MarkupSafe>=2.1.2
python-dotenv>=1.0.0
SQLAlchemy>=2.0.15
typing_extensions>=4.5.0
Werkzeug>=2.3
zipp>=3.15

